Functional tests refactor
- Refactor functional tests to simulate the CCD Response
