+++ src/e2e/resources/appealWithdrawnCallback.json
@@ -39,40 +39,19 @@
       "caseReference": "SC022/14/12423",
       "events": [
         {
-          "id": "e89d246d-8846-4623-821d-d9c38a2aa288",
+          "id": "3182e734-8f89-4ab4-8a5b-e130700811ce",
           "value": {
             "date": "2017-05-24T14:01:18.243",
             "description": "Some Events",
-            "type": "appealCreated"
+            "type": "appealReceived"
           }
         },
         {
-          "id": "2ee8619f-ff94-4c55-b05e-d25e19b3b690",
+          "id": "3bef1aec-f8bd-4212-9290-b228f22f8b4a",
           "value": {
-            "date": "2016-01-16T12:34:56.789",
-            "description": "hearingBooked",
-            "type": "hearingBooked"
-          }
-        }
-      ],
-      "hearings": [
-        {
-          "id": "609523de-ba0a-45e7-8f0b-af3f82ab347c",
-          "value": {
-            "adjourned": null,
-            "hearingDate": "2016-01-01",
-            "time": "12:00",
-            "venue": {
-              "address": {
-                "county": "Aberdeenshire",
-                "line1": "12 The Road Avenue",
-                "line2": "Village",
-                "postcode": "AB12 0HN",
-                "town": "Aberdeen"
-              },
-              "googleMapLink": "http://www.googlemaps.com/aberdeenvenue",
-              "name": "The venue"
-            }
+            "date": "2018-07-25T03:27:04.196",
+            "description": "appealWithdrawn",
+            "type": "appealWithdrawn"
           }
         }
       ],
@@ -97,11 +76,11 @@
     "created_date": [
       2018,
       7,
-      26,
-      11,
+      25,
       15,
-      48,
-      894000000
+      27,
+      2,
+      645000000
     ],
     "data_classification": {
       "appeal": {
@@ -174,7 +153,7 @@
         "classification": "PUBLIC",
         "value": [
           {
-            "id": "e89d246d-8846-4623-821d-d9c38a2aa288",
+            "id": "3182e734-8f89-4ab4-8a5b-e130700811ce",
             "value": {
               "date": "PUBLIC",
               "description": "PUBLIC",
@@ -182,41 +161,11 @@
             }
           },
           {
-            "id": "2ee8619f-ff94-4c55-b05e-d25e19b3b690",
+            "id": "3bef1aec-f8bd-4212-9290-b228f22f8b4a",
             "value": {
               "date": "PUBLIC",
               "description": "PUBLIC",
               "type": "PUBLIC"
-            }
-          }
-        ]
-      },
-      "hearings": {
-        "classification": "PUBLIC",
-        "value": [
-          {
-            "id": "609523de-ba0a-45e7-8f0b-af3f82ab347c",
-            "value": {
-              "adjourned": "PUBLIC",
-              "hearingDate": "PUBLIC",
-              "time": "PUBLIC",
-              "venue": {
-                "classification": "PUBLIC",
-                "value": {
-                  "address": {
-                    "classification": "PUBLIC",
-                    "value": {
-                      "county": "PUBLIC",
-                      "line1": "PUBLIC",
-                      "line2": "PUBLIC",
-                      "postcode": "PUBLIC",
-                      "town": "PUBLIC"
-                    }
-                  },
-                  "googleMapLink": "PUBLIC",
-                  "name": "PUBLIC"
-                }
-              }
             }
           }
         ]
@@ -252,11 +201,11 @@
     "last_modified": [
       2018,
       7,
-      26,
-      11,
+      25,
       15,
-      50,
-      242000000
+      27,
+      3,
+      462000000
     ],
     "security_classification": "PUBLIC",
     "security_classifications": {
@@ -330,7 +279,7 @@
         "classification": "PUBLIC",
         "value": [
           {
-            "id": "e89d246d-8846-4623-821d-d9c38a2aa288",
+            "id": "3182e734-8f89-4ab4-8a5b-e130700811ce",
             "value": {
               "date": "PUBLIC",
               "description": "PUBLIC",
@@ -338,41 +287,11 @@
             }
           },
           {
-            "id": "2ee8619f-ff94-4c55-b05e-d25e19b3b690",
+            "id": "3bef1aec-f8bd-4212-9290-b228f22f8b4a",
             "value": {
               "date": "PUBLIC",
               "description": "PUBLIC",
               "type": "PUBLIC"
-            }
-          }
-        ]
-      },
-      "hearings": {
-        "classification": "PUBLIC",
-        "value": [
-          {
-            "id": "609523de-ba0a-45e7-8f0b-af3f82ab347c",
-            "value": {
-              "adjourned": "PUBLIC",
-              "hearingDate": "PUBLIC",
-              "time": "PUBLIC",
-              "venue": {
-                "classification": "PUBLIC",
-                "value": {
-                  "address": {
-                    "classification": "PUBLIC",
-                    "value": {
-                      "county": "PUBLIC",
-                      "line1": "PUBLIC",
-                      "line2": "PUBLIC",
-                      "postcode": "PUBLIC",
-                      "town": "PUBLIC"
-                    }
-                  },
-                  "googleMapLink": "PUBLIC",
-                  "name": "PUBLIC"
-                }
-              }
             }
           }
         ]
@@ -445,11 +364,11 @@
       "caseReference": "SC022/14/12423",
       "events": [
         {
-          "id": "776d1ead-cb01-43f5-8742-9195bec67399",
+          "id": "42335380-4cf9-4ef4-aeab-9b0471812279",
           "value": {
             "date": "2017-05-24T14:01:18.243",
             "description": "Some Events",
-            "type": "appealCreated"
+            "type": "appealReceived"
           }
         }
       ],
@@ -474,11 +393,11 @@
     "created_date": [
       2018,
       7,
-      26,
-      11,
+      25,
       15,
-      48,
-      894000000
+      27,
+      2,
+      645000000
     ],
     "data_classification": {
       "appeal": {
@@ -551,7 +470,7 @@
         "classification": "PUBLIC",
         "value": [
           {
-            "id": "776d1ead-cb01-43f5-8742-9195bec67399",
+            "id": "42335380-4cf9-4ef4-aeab-9b0471812279",
             "value": {
               "date": "PUBLIC",
               "description": "PUBLIC",
@@ -591,11 +510,11 @@
     "last_modified": [
       2018,
       7,
-      26,
-      11,
+      25,
       15,
-      48,
-      894000000
+      27,
+      2,
+      646000000
     ],
     "security_classification": "PUBLIC",
     "security_classifications": {
@@ -669,7 +588,7 @@
         "classification": "PUBLIC",
         "value": [
           {
-            "id": "776d1ead-cb01-43f5-8742-9195bec67399",
+            "id": "42335380-4cf9-4ef4-aeab-9b0471812279",
             "value": {
               "date": "PUBLIC",
               "description": "PUBLIC",
@@ -706,5 +625,5 @@
     },
     "state": "appealCreated"
   },
-  "event_id": "hearingBooked"
+  "event_id": "appealWithdrawn"
 }
