+++ src/e2e/resources/hearingAdjournedCallback.json
@@ -39,40 +39,11 @@
       "caseReference": "SC022/14/12423",
       "events": [
         {
-          "id": "e89d246d-8846-4623-821d-d9c38a2aa288",
+          "id": "96f908a4-c05f-454e-a024-94c115cd3d47",
           "value": {
             "date": "2017-05-24T14:01:18.243",
             "description": "Some Events",
-            "type": "appealCreated"
-          }
-        },
-        {
-          "id": "2ee8619f-ff94-4c55-b05e-d25e19b3b690",
-          "value": {
-            "date": "2016-01-16T12:34:56.789",
-            "description": "hearingBooked",
-            "type": "hearingBooked"
-          }
-        }
-      ],
-      "hearings": [
-        {
-          "id": "609523de-ba0a-45e7-8f0b-af3f82ab347c",
-          "value": {
-            "adjourned": null,
-            "hearingDate": "2016-01-01",
-            "time": "12:00",
-            "venue": {
-              "address": {
-                "county": "Aberdeenshire",
-                "line1": "12 The Road Avenue",
-                "line2": "Village",
-                "postcode": "AB12 0HN",
-                "town": "Aberdeen"
-              },
-              "googleMapLink": "http://www.googlemaps.com/aberdeenvenue",
-              "name": "The venue"
-            }
+            "type": "appealReceived"
           }
         }
       ],
@@ -97,11 +68,11 @@
     "created_date": [
       2018,
       7,
-      26,
-      11,
+      25,
       15,
-      48,
-      894000000
+      9,
+      17,
+      242000000
     ],
     "data_classification": {
       "appeal": {
@@ -174,49 +145,11 @@
         "classification": "PUBLIC",
         "value": [
           {
-            "id": "e89d246d-8846-4623-821d-d9c38a2aa288",
+            "id": "96f908a4-c05f-454e-a024-94c115cd3d47",
             "value": {
               "date": "PUBLIC",
               "description": "PUBLIC",
               "type": "PUBLIC"
-            }
-          },
-          {
-            "id": "2ee8619f-ff94-4c55-b05e-d25e19b3b690",
-            "value": {
-              "date": "PUBLIC",
-              "description": "PUBLIC",
-              "type": "PUBLIC"
-            }
-          }
-        ]
-      },
-      "hearings": {
-        "classification": "PUBLIC",
-        "value": [
-          {
-            "id": "609523de-ba0a-45e7-8f0b-af3f82ab347c",
-            "value": {
-              "adjourned": "PUBLIC",
-              "hearingDate": "PUBLIC",
-              "time": "PUBLIC",
-              "venue": {
-                "classification": "PUBLIC",
-                "value": {
-                  "address": {
-                    "classification": "PUBLIC",
-                    "value": {
-                      "county": "PUBLIC",
-                      "line1": "PUBLIC",
-                      "line2": "PUBLIC",
-                      "postcode": "PUBLIC",
-                      "town": "PUBLIC"
-                    }
-                  },
-                  "googleMapLink": "PUBLIC",
-                  "name": "PUBLIC"
-                }
-              }
             }
           }
         ]
@@ -252,11 +185,11 @@
     "last_modified": [
       2018,
       7,
-      26,
-      11,
+      25,
       15,
-      50,
-      242000000
+      9,
+      18,
+      367000000
     ],
     "security_classification": "PUBLIC",
     "security_classifications": {
@@ -330,49 +263,11 @@
         "classification": "PUBLIC",
         "value": [
           {
-            "id": "e89d246d-8846-4623-821d-d9c38a2aa288",
+            "id": "96f908a4-c05f-454e-a024-94c115cd3d47",
             "value": {
               "date": "PUBLIC",
               "description": "PUBLIC",
               "type": "PUBLIC"
-            }
-          },
-          {
-            "id": "2ee8619f-ff94-4c55-b05e-d25e19b3b690",
-            "value": {
-              "date": "PUBLIC",
-              "description": "PUBLIC",
-              "type": "PUBLIC"
-            }
-          }
-        ]
-      },
-      "hearings": {
-        "classification": "PUBLIC",
-        "value": [
-          {
-            "id": "609523de-ba0a-45e7-8f0b-af3f82ab347c",
-            "value": {
-              "adjourned": "PUBLIC",
-              "hearingDate": "PUBLIC",
-              "time": "PUBLIC",
-              "venue": {
-                "classification": "PUBLIC",
-                "value": {
-                  "address": {
-                    "classification": "PUBLIC",
-                    "value": {
-                      "county": "PUBLIC",
-                      "line1": "PUBLIC",
-                      "line2": "PUBLIC",
-                      "postcode": "PUBLIC",
-                      "town": "PUBLIC"
-                    }
-                  },
-                  "googleMapLink": "PUBLIC",
-                  "name": "PUBLIC"
-                }
-              }
             }
           }
         ]
@@ -445,11 +340,11 @@
       "caseReference": "SC022/14/12423",
       "events": [
         {
-          "id": "776d1ead-cb01-43f5-8742-9195bec67399",
+          "id": "03e2955a-6f6b-4e13-8bc8-1b782641d8fc",
           "value": {
             "date": "2017-05-24T14:01:18.243",
             "description": "Some Events",
-            "type": "appealCreated"
+            "type": "appealReceived"
           }
         }
       ],
@@ -474,11 +369,11 @@
     "created_date": [
       2018,
       7,
-      26,
-      11,
+      25,
       15,
-      48,
-      894000000
+      9,
+      17,
+      242000000
     ],
     "data_classification": {
       "appeal": {
@@ -551,7 +446,7 @@
         "classification": "PUBLIC",
         "value": [
           {
-            "id": "776d1ead-cb01-43f5-8742-9195bec67399",
+            "id": "03e2955a-6f6b-4e13-8bc8-1b782641d8fc",
             "value": {
               "date": "PUBLIC",
               "description": "PUBLIC",
@@ -591,11 +486,11 @@
     "last_modified": [
       2018,
       7,
-      26,
-      11,
+      25,
       15,
-      48,
-      894000000
+      9,
+      17,
+      243000000
     ],
     "security_classification": "PUBLIC",
     "security_classifications": {
@@ -669,7 +564,7 @@
         "classification": "PUBLIC",
         "value": [
           {
-            "id": "776d1ead-cb01-43f5-8742-9195bec67399",
+            "id": "03e2955a-6f6b-4e13-8bc8-1b782641d8fc",
             "value": {
               "date": "PUBLIC",
               "description": "PUBLIC",
@@ -706,5 +601,5 @@
     },
     "state": "appealCreated"
   },
-  "event_id": "hearingBooked"
+  "event_id": "hearingAdjourned"
 }
