+++ src/main/resources/application.yaml
@@ -65,6 +65,7 @@ gov.uk.notification.api.key: ${NOTIFICATION_API_KEY:emailnotificationkey}
 gov.uk.notification.api.testKey: ${NOTIFICATION_API_TEST_KEY:emailnotificationtestkey}
 gov.uk.notification.api.testRecipients:
   - "sscstest+notify@greencroftconsulting.com"
+  - "sscstest+notify2@greencroftconsulting.com"
   - "07398785050"
 
 subscription.hmac.secret.text: ${EMAIL_MAC_SECRET_TEXT:our-big-secret}
